--- npm-debug.log
@@ -1,49 +1,0 @@
-0 info it worked if it ends with ok
-1 verbose cli [ 'C:\\Program Files\\nodejs\\node.exe',
-1 verbose cli   'C:\\Program Files\\nodejs\\node_modules\\npm\\bin\\npm-cli.js',
-1 verbose cli   'run',
-1 verbose cli   'dev' ]
-2 info using npm@3.10.10
-3 info using node@v6.10.2
-4 verbose run-script [ 'predev', 'dev', 'postdev' ]
-5 info lifecycle react_simple@1.0.0~predev: react_simple@1.0.0
-6 silly lifecycle react_simple@1.0.0~predev: no script for predev, continuing
-7 info lifecycle react_simple@1.0.0~dev: react_simple@1.0.0
-8 verbose lifecycle react_simple@1.0.0~dev: unsafe-perm in lifecycle true
-9 verbose lifecycle react_simple@1.0.0~dev: PATH: C:\Program Files\nodejs\node_modules\npm\bin\node-gyp-bin;C:\xampp\htdocs\react\airlines-react-redux\node_modules\.bin;C:\Program Files\Git\mingw64\bin;C:\Program Files\Git\usr\bin;C:\Users\Denis.BykovDenis\bin;C:\Program Files\Docker\Docker\Resources\bin;C:\Windows\system32;C:\Windows;C:\Windows\System32\Wbem;C:\Windows\System32\WindowsPowerShell\v1.0;C:\php;C:\ProgramData\ComposerSetup\bin;C:\Users\Denis.BykovDenis\AppData\Roaming\npm;C:\Program Files (x86)\NVIDIA Corporation\PhysX\Common;C:\Program Files\Git;C:\Python27\Scripts;C:\WINDOWS\system32;C:\WINDOWS;C:\WINDOWS\System32\Wbem;C:\WINDOWS\System32\WindowsPowerShell\v1.0;E:\WEB software\автотестирование\phantomjs-2.1.1-windows\phantomjs-2.1.1-windows\bin;C:\Program Files (x86)\Skype\Phone;C:\Program Files\nodejs;C:\Ruby23-x64\bin;C:\Users\Denis.BykovDenis\AppData\Local\Microsoft\WindowsApps;C:\Users\Denis.BykovDenis\AppData\Roaming\npm;C:\Users\Denis.BykovDenis\AppData\Local\atom\bin
-10 verbose lifecycle react_simple@1.0.0~dev: CWD: C:\xampp\htdocs\react\airlines-react-redux
-11 silly lifecycle react_simple@1.0.0~dev: Args: [ '/d /s /c',
-11 silly lifecycle   'REFRESH=refresh webpack-dev-server --inline --colors --hot --config webpack.config.js' ]
-12 silly lifecycle react_simple@1.0.0~dev: Returned: code: 1  signal: null
-13 info lifecycle react_simple@1.0.0~dev: Failed to exec dev script
-14 verbose stack Error: react_simple@1.0.0 dev: `REFRESH=refresh webpack-dev-server --inline --colors --hot --config webpack.config.js`
-14 verbose stack Exit status 1
-14 verbose stack     at EventEmitter.<anonymous> (C:\Program Files\nodejs\node_modules\npm\lib\utils\lifecycle.js:255:16)
-14 verbose stack     at emitTwo (events.js:106:13)
-14 verbose stack     at EventEmitter.emit (events.js:191:7)
-14 verbose stack     at ChildProcess.<anonymous> (C:\Program Files\nodejs\node_modules\npm\lib\utils\spawn.js:40:14)
-14 verbose stack     at emitTwo (events.js:106:13)
-14 verbose stack     at ChildProcess.emit (events.js:191:7)
-14 verbose stack     at maybeClose (internal/child_process.js:886:16)
-14 verbose stack     at Process.ChildProcess._handle.onexit (internal/child_process.js:226:5)
-15 verbose pkgid react_simple@1.0.0
-16 verbose cwd C:\xampp\htdocs\react\airlines-react-redux
-17 error Windows_NT 10.0.14393
-18 error argv "C:\\Program Files\\nodejs\\node.exe" "C:\\Program Files\\nodejs\\node_modules\\npm\\bin\\npm-cli.js" "run" "dev"
-19 error node v6.10.2
-20 error npm  v3.10.10
-21 error code ELIFECYCLE
-22 error react_simple@1.0.0 dev: `REFRESH=refresh webpack-dev-server --inline --colors --hot --config webpack.config.js`
-22 error Exit status 1
-23 error Failed at the react_simple@1.0.0 dev script 'REFRESH=refresh webpack-dev-server --inline --colors --hot --config webpack.config.js'.
-23 error Make sure you have the latest version of node.js and npm installed.
-23 error If you do, this is most likely a problem with the react_simple package,
-23 error not with npm itself.
-23 error Tell the author that this fails on your system:
-23 error     REFRESH=refresh webpack-dev-server --inline --colors --hot --config webpack.config.js
-23 error You can get information on how to open an issue for this project with:
-23 error     npm bugs react_simple
-23 error Or if that isn't available, you can get their info via:
-23 error     npm owner ls react_simple
-23 error There is likely additional logging output above.
-24 verbose exit [ 1, true ]
