+++ package.json
@@ -82,6 +82,7 @@
     "react": "^15.5.4",
     "react-dom": "^15.5.4",
     "react-redux": "^4.0.0",
+    "react-router": "^4.1.1",
     "redux": "^3.0.4",
     "redux-devtools-extension": "^2.13.0",
     "redux-thunk": "^2.1.0",
