+++ src/js/containers/airptorts.jsx
@@ -1,0 +1,40 @@
+/**
+ * Created by Denis on 08.06.2017.
+ */
+import React, { Component } from 'react';
+import PropTypes from 'prop-types';
+import { connect } from 'react-redux';
+// модуль комбинирования работы нескольких актшионов
+import { bindActionCreators } from 'redux';
+import * as getActionData from '../redux/actions/actions';
+// подгрузка компонентов
+import DataTableAirports from '../components/airports/';
+
+class Airports extends Component {
+  static get propTypes() {
+    return {
+      getData: PropTypes.object.isRequired,
+      currentStore: PropTypes.object.isRequired,
+    };
+  }
+  componentDidMount() {
+    this.props.getData.getAirportsData();
+  }
+  render() {
+    const airports = this.props.currentStore.Reducer.airports;
+    return (
+      <div>
+        <DataTableAirports infoData={airports} />
+      </div>
+    );
+  }
+}
+
+const getDataProps = dispatch => (
+  { getData: bindActionCreators(getActionData, dispatch) }
+);
+
+export default connect(
+  state => ({ currentStore: state }),
+  getDataProps
+)(Airports);
