+++ src/js/containers/arrival.jsx
@@ -1,0 +1,37 @@
+import React, { Component } from 'react';
+import PropTypes from 'prop-types';
+import { connect } from 'react-redux';
+// модуль комбинирования работы нескольких актшионов
+import { bindActionCreators } from 'redux';
+import * as getActionData from '../redux/actions/actions';
+// подгрузка компонентов
+import DataTableSheduleArrival from '../components/shedule-arrival/';
+
+class SheduleArrival extends Component {
+  static get propTypes() {
+    return {
+      getData: PropTypes.object.isRequired,
+      currentStore: PropTypes.object.isRequired,
+    };
+  }
+  componentDidMount() {
+    this.props.getData.getArrivingData();
+  }
+  render() {
+    const arriving = this.props.currentStore.Reducer.shedules.arriving;
+    return (
+      <div>
+        <DataTableSheduleArrival infoData={arriving} />
+      </div>
+    );
+  }
+}
+
+const getDataProps = dispatch => (
+  { getData: bindActionCreators(getActionData, dispatch) }
+);
+
+export default connect(
+  state => ({ currentStore: state }),
+  getDataProps
+)(SheduleArrival);
