+++ src/js/containers/departure.jsx
@@ -1,0 +1,38 @@
+import React, { Component } from 'react';
+import PropTypes from 'prop-types';
+import { connect } from 'react-redux';
+// модуль комбинирования работы нескольких актшионов
+import { bindActionCreators } from 'redux';
+import * as getActionData from '../redux/actions/actions';
+// подгрузка компонентов
+import SheduleDeparture from '../components/shedule-departure/';
+
+
+class Departure extends Component {
+  static get propTypes() {
+    return {
+      getData: PropTypes.object.isRequired,
+      currentStore: PropTypes.object.isRequired,
+    };
+  }
+  componentDidMount() {
+    this.props.getData.getDepartingData();
+  }
+  render() {
+    const departing = this.props.currentStore.Reducer.shedules.departing;
+    return (
+      <div>
+        <SheduleDeparture infoData={departing} />
+      </div>
+    );
+  }
+}
+
+const getDataProps = dispatch => (
+  { getData: bindActionCreators(getActionData, dispatch) }
+);
+
+export default connect(
+  state => ({ currentStore: state }),
+  getDataProps
+)(Departure);
